requirements.txt
apache-airflow==2.7.1
beautifulsoup4==4.12.2
joblib==1.3.2
numpy
pandas
Requests==2.31.0
scikit_learn==1.3.0
streamlit==1.26.0
xgboost==2.0.0
